+++ frontend/src/pages/Login.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { useAuth } from '../context/AuthContext';
-import Button from '../components/Button';
-import Input from '../components/Input';
+import Button from '../components/common/Button';
+import Input from '../components/common/Input';
 
 const Login = () => {
   const [formData, setFormData] = useState({
